+++ fuel-vm/src/storage/memory.rs
@@ -3,6 +3,7 @@ use crate::storage::{
     ContractsAssets,
     ContractsRawCode,
     ContractsState,
+    ContractsStateData,
     ContractsStateKey,
     InterpreterStorage,
 };
@@ -39,7 +40,7 @@ use super::interpreter::ContractsAssetsStorage;
 struct MemoryStorageInner {
     contracts: BTreeMap<ContractId, Contract>,
     balances: BTreeMap<ContractsAssetKey, Word>,
-    contract_state: BTreeMap<ContractsStateKey, Bytes32>,
+    contract_state: BTreeMap<ContractsStateKey, ContractsStateData>,
 }
 
 #[derive(Debug, Clone)]
@@ -73,7 +74,7 @@ impl MemoryStorage {
     /// Iterate over all contract state in storage
     pub fn all_contract_state(
         &self,
-    ) -> impl Iterator<Item = (&ContractsStateKey, &Bytes32)> {
+    ) -> impl Iterator<Item = (&ContractsStateKey, &ContractsStateData)> {
         self.memory.contract_state.iter()
     }
 
@@ -82,13 +83,11 @@ impl MemoryStorage {
         &self,
         contract: &ContractId,
         key: &Bytes32,
-    ) -> Cow<'_, Bytes32> {
-        const DEFAULT_STATE: Bytes32 = Bytes32::zeroed();
-
+    ) -> Cow<'_, ContractsStateData> {
         self.storage::<ContractsState>()
             .get(&(contract, key).into())
             .expect("Infallible")
-            .unwrap_or(Cow::Borrowed(&DEFAULT_STATE))
+            .unwrap_or(Cow::Owned(ContractsStateData::default()))
     }
 
     /// Set the transacted state to the memory state.
@@ -156,7 +155,7 @@ impl StorageMutate<ContractsRawCode> for MemoryStorage {
 }
 
 impl StorageWrite<ContractsRawCode> for MemoryStorage {
-    fn write(&mut self, key: &ContractId, buf: Vec<u8>) -> Result<usize, Infallible> {
+    fn write(&mut self, key: &ContractId, buf: &[u8]) -> Result<usize, Infallible> {
         let size = buf.len();
         self.memory.contracts.insert(*key, Contract::from(buf));
         Ok(size)
@@ -164,22 +163,21 @@ impl StorageWrite<ContractsRawCode> for MemoryStorage {
 
     fn replace(
         &mut self,
-        key: &<ContractsRawCode as Mappable>::Key,
-        buf: Vec<u8>,
-    ) -> Result<(usize, Option<Vec<u8>>), Self::Error>
-    where
-        Self: StorageSize<ContractsRawCode>,
-    {
+        key: &ContractId,
+        buf: &[u8],
+    ) -> Result<(usize, Option<Vec<u8>>), Self::Error> {
         let size = buf.len();
-        let last = self.memory.contracts.insert(*key, Contract::from(buf));
-        Ok((size, last.map(Vec::from)))
+        let prev = self
+            .memory
+            .contracts
+            .insert(*key, Contract::from(buf))
+            .map(Into::into);
+        Ok((size, prev))
     }
 
-    fn take(
-        &mut self,
-        key: &<ContractsRawCode as Mappable>::Key,
-    ) -> Result<Option<Vec<u8>>, Self::Error> {
-        Ok(self.memory.contracts.remove(key).map(Vec::from))
+    fn take(&mut self, key: &ContractId) -> Result<Option<Vec<u8>>, Self::Error> {
+        let prev = self.memory.contracts.remove(key).map(Into::into);
+        Ok(prev)
     }
 }
 
@@ -248,7 +246,8 @@ impl StorageInspect<ContractsState> for MemoryStorage {
     fn get(
         &self,
         key: &<ContractsState as Mappable>::Key,
-    ) -> Result<Option<Cow<'_, Bytes32>>, Infallible> {
+    ) -> Result<Option<Cow<'_, <ContractsState as Mappable>::OwnedValue>>, Infallible>
+    {
         Ok(self.memory.contract_state.get(key).map(Cow::Borrowed))
     }
 
@@ -264,16 +263,93 @@ impl StorageMutate<ContractsState> for MemoryStorage {
     fn insert(
         &mut self,
         key: &<ContractsState as Mappable>::Key,
-        value: &Bytes32,
-    ) -> Result<Option<Bytes32>, Infallible> {
-        Ok(self.memory.contract_state.insert(*key, *value))
+        value: &<ContractsState as Mappable>::Value,
+    ) -> Result<Option<<ContractsState as Mappable>::OwnedValue>, Infallible> {
+        Ok(self.memory.contract_state.insert(*key, value.into()))
     }
 
     fn remove(
         &mut self,
         key: &<ContractsState as Mappable>::Key,
-    ) -> Result<Option<Bytes32>, Infallible> {
+    ) -> Result<Option<ContractsStateData>, Infallible> {
         Ok(self.memory.contract_state.remove(key))
+    }
+}
+
+impl StorageWrite<ContractsState> for MemoryStorage {
+    fn write(
+        &mut self,
+        key: &<ContractsState as Mappable>::Key,
+        buf: &[u8],
+    ) -> Result<usize, Infallible> {
+        let size = buf.len();
+        self.memory
+            .contract_state
+            .insert(*key, ContractsStateData::from(buf));
+        Ok(size)
+    }
+
+    fn replace(
+        &mut self,
+        key: &<ContractsState as Mappable>::Key,
+        buf: &[u8],
+    ) -> Result<(usize, Option<Vec<u8>>), Self::Error>
+    where
+        Self: StorageSize<ContractsState>,
+    {
+        let size = buf.len();
+        let prev = self
+            .memory
+            .contract_state
+            .insert(*key, ContractsStateData::from(buf))
+            .map(Into::into);
+        Ok((size, prev))
+    }
+
+    fn take(
+        &mut self,
+        key: &<ContractsState as Mappable>::Key,
+    ) -> Result<Option<Vec<u8>>, Self::Error> {
+        let prev = self.memory.contract_state.remove(key).map(Into::into);
+        Ok(prev)
+    }
+}
+
+impl StorageSize<ContractsState> for MemoryStorage {
+    fn size_of_value(
+        &self,
+        key: &<ContractsState as Mappable>::Key,
+    ) -> Result<Option<usize>, Infallible> {
+        Ok(self
+            .memory
+            .contract_state
+            .get(key)
+            .map(|c| c.as_ref().len()))
+    }
+}
+
+impl StorageRead<ContractsState> for MemoryStorage {
+    fn read(
+        &self,
+        key: &<ContractsState as Mappable>::Key,
+        buf: &mut [u8],
+    ) -> Result<Option<usize>, Self::Error> {
+        Ok(self.memory.contract_state.get(key).map(|data| {
+            let len = buf.len().min(data.as_ref().len());
+            buf.copy_from_slice(&data.as_ref()[..len]);
+            len
+        }))
+    }
+
+    fn read_alloc(
+        &self,
+        key: &<ContractsState as Mappable>::Key,
+    ) -> Result<Option<Vec<u8>>, Self::Error> {
+        Ok(self
+            .memory
+            .contract_state
+            .get(key)
+            .map(|c| c.as_ref().to_vec()))
     }
 }
 
@@ -306,7 +382,7 @@ impl InterpreterStorage for MemoryStorage {
         id: &ContractId,
         start_key: &Bytes32,
         range: usize,
-    ) -> Result<Vec<Option<Cow<Bytes32>>>, Self::DataError> {
+    ) -> Result<Vec<Option<Cow<ContractsStateData>>>, Self::DataError> {
         let start: ContractsStateKey = (id, start_key).into();
         let end: ContractsStateKey = (id, &Bytes32::new([u8::MAX; 32])).into();
         let mut iter = self.memory.contract_state.range(start..end);
@@ -338,14 +414,19 @@ impl InterpreterStorage for MemoryStorage {
         .collect())
     }
 
-    fn contract_state_insert_range(
+    fn contract_state_insert_range<'a, I>(
         &mut self,
         contract: &ContractId,
         start_key: &Bytes32,
-        values: &[Bytes32],
-    ) -> Result<usize, Self::DataError> {
+        values: I,
+    ) -> Result<usize, Self::DataError>
+    where
+        I: Iterator<Item = &'a [u8]>,
+    {
+        let storage: &mut dyn StorageWrite<ContractsState, Error = Self::DataError> =
+            self;
         let mut unset_count = 0;
-        let values: Vec<_> = core::iter::successors(Some(**start_key), |n| {
+        core::iter::successors(Some(**start_key), |n| {
             let mut n = *n;
             if add_one(&mut n) {
                 None
@@ -354,15 +435,14 @@ impl InterpreterStorage for MemoryStorage {
             }
         })
         .zip(values)
-        .map(|(key, value)| {
-            let key = (contract, &Bytes32::from(key)).into();
-            if !self.memory.contract_state.contains_key(&key) {
+        .try_for_each(|(key, value)| {
+            let key: ContractsStateKey = (contract, &Bytes32::from(key)).into();
+            if !storage.contains_key(&key)? {
                 unset_count += 1;
             }
-            (key, *value)
-        })
-        .collect();
-        self.memory.contract_state.extend(values);
+            storage.write(&key, value)?;
+            Ok::<_, Self::DataError>(())
+        })?;
         Ok(unset_count)
     }
 
@@ -421,23 +501,23 @@ mod tests {
         ]
     }
 
-    #[test_case(&[&[0u8; 32]], &[0u8; 32], 1 => vec![Some(Bytes32::zeroed())])]
-    #[test_case(&[&[0u8; 32]], &[0u8; 32], 0 => Vec::<Option<Bytes32>>::with_capacity(0))]
+    #[test_case(&[&[0u8; 32]], &[0u8; 32], 1 => vec![Some(Default::default())])]
+    #[test_case(&[&[0u8; 32]], &[0u8; 32], 0 => Vec::<Option<ContractsStateData>>::with_capacity(0))]
     #[test_case(&[], &[0u8; 32], 1 => vec![None])]
     #[test_case(&[], &[1u8; 32], 1 => vec![None])]
     #[test_case(&[&[0u8; 32]], &key(1), 2 => vec![None, None])]
-    #[test_case(&[&key(1), &key(3)], &[0u8; 32], 4 => vec![None, Some(Bytes32::zeroed()), None, Some(Bytes32::zeroed())])]
-    #[test_case(&[&[0u8; 32], &key(1)], &[0u8; 32], 1 => vec![Some(Bytes32::zeroed())])]
+    #[test_case(&[&key(1), &key(3)], &[0u8; 32], 4 => vec![None, Some(Default::default()), None, Some(Default::default())])]
+    #[test_case(&[&[0u8; 32], &key(1)], &[0u8; 32], 1 => vec![Some(Default::default())])]
     fn test_contract_state_range(
         store: &[&[u8; 32]],
         start: &[u8; 32],
         range: usize,
-    ) -> Vec<Option<Bytes32>> {
+    ) -> Vec<Option<ContractsStateData>> {
         let mut mem = MemoryStorage::default();
         for k in store {
             mem.memory.contract_state.insert(
                 (&ContractId::default(), &(**k).into()).into(),
-                Bytes32::zeroed(),
+                Default::default(),
             );
         }
         mem.contract_state_range(&ContractId::default(), &(*start).into(), range)